functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.11.2
pandas==1.4.2
azure-kusto-data==3.1.3
azure-kusto-ingest==3.1.3
